flaskAuth/requirements.txt
pyrebase4==4.7.1
flask==3.0.1
python-dotenv==1.0.1
googleapis-common-protos==1.62.0
flask-sqlalchemy
flask-bcrypt
flask-session
redis
flask-cors
